requirements.txt
Flask
selenium
requests
transformers
tensorflow-cpu
beautifulsoup4
pandas
numpy
gunicorn